+++ src/Symfony/Component/HttpKernel/CacheWarmer/CacheWarmerAggregate.php
@@ -11,6 +11,10 @@
 
 namespace Symfony\Component\HttpKernel\CacheWarmer;
 
+use Symfony\Component\HttpKernel\KernelInterface;
+use Symfony\Component\Finder\Finder;
+use Symfony\Component\Filesystem\Filesystem;
+
 /**
  * Aggregates several cache warmers into a single one.
  *
@@ -18,8 +22,12 @@ namespace Symfony\Component\HttpKernel\CacheWarmer;
  */
 class CacheWarmerAggregate implements CacheWarmerInterface
 {
+    const NEW_CACHE_FOLDER_SUFFIX = '_new';
+    const OLD_CACHE_FOLDER_SUFFIX = '_old';
+
     protected $warmers;
     protected $optionalsEnabled;
+    protected $kernelSuffix;
 
     public function __construct(array $warmers = array())
     {
@@ -27,9 +35,12 @@ class CacheWarmerAggregate implements CacheWarmerInterface
         $this->optionalsEnabled = false;
     }
 
-    public function enableOptionalWarmers()
+    /**
+     * @param Boolean $enabled
+     */
+    public function enableOptionalWarmers($enabled = true)
     {
-        $this->optionalsEnabled = true;
+        $this->optionalsEnabled = (Boolean) $enabled;
     }
 
     /**
@@ -69,5 +80,119 @@ class CacheWarmerAggregate implements CacheWarmerInterface
     public function add(CacheWarmerInterface $warmer)
     {
         $this->warmers[] = $warmer;
+    }
+
+    /**
+     * Warms-up a kernel according to the given configuration.
+     *
+     * @param KernelInterface  $kernel           The kernel
+     * @param Filesystem       $fs               The filesystem
+     * @param string           $cacheDir         The cache directory where the cache files will be generated
+     * @param string           $env              The environment (ie 'prod', 'dev')
+     * @param Boolean          $debug            The value of the %kernel.debug% parameter
+     * @param Boolean          $optionalsEnabled Whether the optional warmers should be enabled
+     * @param string|null      $tempCacheDir     A temporary working directory
+     */
+    public function warmUpForEnv(KernelInterface $kernel, FileSystem $fs, $cacheDir, $env, $debug, $optionalsEnabled, $tempCacheDir = null)
+    {
+        if (null === $tempCacheDir) {
+            $tempCacheDir = rtrim($cacheDir, '/\\') . self::NEW_CACHE_FOLDER_SUFFIX;
+        }
+
+        $fs->remove($tempCacheDir);
+
+        $class = get_class($kernel);
+        $namespace = '';
+        if (false !== $pos = strrpos($class, '\\')) {
+            $namespace = substr($class, 0, $pos);
+            $class = substr($class, $pos + 1);
+        }
+
+        $tempKernel = $this->getTempKernel($kernel, $namespace, $class, $tempCacheDir, $env, $debug, $fs);
+        $tempKernel->boot();
+
+        $warmer = $tempKernel->getContainer()->get('cache_warmer');
+        $warmer->enableOptionalWarmers($optionalsEnabled);
+        $warmer->warmUp($tempCacheDir);
+
+        // fix container files and classes
+        $regex = '/' . preg_quote($this->getTempKernelSuffix(), '/') . '/';
+        foreach (Finder::create()->files()->name(get_class($tempKernel->getContainer()) . '*')->in($tempCacheDir) as $file) {
+            $content = preg_replace($regex, '', $file->getContents());
+
+            // fix absolute paths to the cache directory
+            $content = preg_replace(
+                '/'.preg_quote($tempCacheDir, '/').'/',
+                preg_replace('/' . self::NEW_CACHE_FOLDER_SUFFIX . '$/', '', $tempCacheDir),
+                $content
+            );
+
+            file_put_contents(preg_replace($regex, '', $file), $content);
+            $fs->remove($file);
+        }
+
+        $from = '/C:\d+:"' . preg_quote($class . $this->getTempKernelSuffix(), '/').'"/';
+        $to = sprintf('C:%d:"%s"', strlen($class), $class);
+
+        foreach (Finder::create()->files()->name('*.meta')->in($tempCacheDir) as $file) {
+            // Fix references to the kernel
+            $file->putContents(preg_replace($from, $to, $file->getContents()));
+        }
+
+        $oldCacheDir = rtrim($cacheDir, '/\\') . self::OLD_CACHE_FOLDER_SUFFIX;
+        $fs
+            ->rename($cacheDir, $oldCacheDir)
+            ->rename($tempCacheDir, $cacheDir)
+            ->remove($oldCacheDir)
+        ;
+    }
+
+    /**
+     * @return string A unique suffix used to identify the temporary kernel
+     */
+    protected function getTempKernelSuffix()
+    {
+        if (null === $this->kernelSuffix) {
+            $this->kernelSuffix = '__' . uniqid() . '__';
+        }
+
+        return $this->kernelSuffix;
+    }
+
+    protected function getTempKernel(KernelInterface $parent, $namespace, $class, $warmupDir, $env, $debug, FileSystem $fs)
+    {
+        $suffix = $this->getTempKernelSuffix();
+        $rootDir = $parent->getRootDir();
+        $code = <<<EOF
+<?php
+
+namespace $namespace
+{
+    class $class$suffix extends $class
+    {
+        public function getCacheDir()
+        {
+            return '$warmupDir';
+        }
+
+        public function getRootDir()
+        {
+            return '$rootDir';
+        }
+
+        protected function getContainerClass()
+        {
+            return parent::getContainerClass().'$suffix';
+        }
+    }
+}
+EOF;
+        $fs->mkdir($warmupDir);
+        file_put_contents($file = $warmupDir.'/kernel.tmp', $code);
+        require_once $file;
+        $fs->remove($file);
+        $class = "$namespace\\$class$suffix";
+
+        return new $class($env, $debug);
     }
 }